refactor tooltip
